Move PreRelease class to its own file
